+++ check/legacy_guide_file_test.go
@@ -30,6 +30,12 @@ func TestLegacyGuideFileCheck(t *testing.T) {
 			ExpectError: true,
 		},
 		{
+			Name:        "invalid frontmatter with sidebar_current",
+			BasePath:    "testdata/invalid-legacy-files",
+			Path:        "guide_with_sidebar_current.html.markdown",
+			ExpectError: true,
+		},
+		{
 			Name:        "invalid frontmatter without layout",
 			BasePath:    "testdata/invalid-legacy-files",
 			Path:        "guide_without_layout.html.markdown",