+++ CHANGELOG.md
@@ -4,6 +4,8 @@ ENHANCEMENTS
 
 * check: Verify number of documentation files for Terraform Registry storage limits
 * check: Verify size of documentation files for Terraform Registry storage limits
+* check: Verify all known data sources and resources have an associated documentation file (if `-providers-schema-json` is provided)
+* check: Verify no extraneous or incorrectly named documentation files exist (if `-providers-schema-json` is provided)
 
 # v0.1.2